Seperated pass and plugin registration, moved pass to tooling common for us to be able to have tests on it
