+++ src/lib/Plugins/LoadHIPFATBinaryPlugin/LoadHIPFATBinaryInfoPlugin.cpp
@@ -1,0 +1,56 @@
+//===-- LoadHIPFATBinaryInfo.cpp --------------------------------*- C++ -*-===//
+// Copyright 2026 @ Northeastern University Computer Architecture Lab
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//===----------------------------------------------------------------------===//
+/// This registers the HIPFATBinaryInfoPass as a plugin. We seperate the pass
+/// and the plugin so that the pass can be used independendently for testing
+//===----------------------------------------------------------------------===//
+#include "luthier/Tooling/LoadHIPFATBinaryInfoPass.h"
+#include <llvm/Analysis/ValueTracking.h>
+#include <llvm/IR/LLVMContext.h>
+#include <llvm/IR/Module.h>
+#include <llvm/Passes/PassBuilder.h>
+#include <llvm/Plugins/PassPlugin.h>
+#include <llvm/Support/FormatVariadic.h>
+
+#undef DEBUG_TYPE
+#define DEBUG_TYPE "luthier-load-hip-fat-binary-info-pass"
+
+llvm::PassPluginLibraryInfo getLuthierHIPFATBinaryPluginInfo() {
+  return {LLVM_PLUGIN_API_VERSION, DEBUG_TYPE, LLVM_VERSION_STRING,
+          [](llvm::PassBuilder &PB) {
+            PB.registerPipelineStartEPCallback(
+                [](llvm::ModulePassManager &MPM, llvm::OptimizationLevel Opt) {
+                  MPM.addPass(luthier::LoadHIPFATBinaryInfoPass());
+                });
+            /// We register a Pipeline Parsing Callback so we can invoke opt on
+            /// this pass
+            PB.registerPipelineParsingCallback(
+                [](llvm::StringRef Name, llvm::ModulePassManager &MPM,
+                   llvm::ArrayRef<llvm::PassBuilder::PipelineElement>) {
+                  if (Name == DEBUG_TYPE) {
+                    MPM.addPass(luthier::LoadHIPFATBinaryInfoPass());
+                    return true;
+                  }
+                  return false;
+                });
+          }};
+}
+
+#ifndef LLVM_LUTHIERHIPFATBINARYPLUGIN_LINK_INTO_TOOLS
+extern "C" LLVM_ATTRIBUTE_WEAK ::llvm::PassPluginLibraryInfo
+llvmGetPassPluginInfo() {
+  return getLuthierHIPFATBinaryPluginInfo();
+}
+#endif
