+++ src/lib/Plugins/OptPlugin/OptPlugin.cpp
@@ -24,10 +24,10 @@
 #include "luthier/Tooling/InitialEntryPointAnalysis.h"
 #include "luthier/Tooling/InstructionTracesAnalysis.h"
 // #include "luthier/Tooling/InstrumentationPMDriver.h"
-#include "luthier/Tooling/IntrinsicMIRLoweringPass.h"
+//#include "luthier/Tooling/IntrinsicMIRLoweringPass.h"
 // #include "luthier/Tooling/LRCallgraph.h"
 // #include "luthier/Tooling/MMISlotIndexesAnalysis.h"
-#include "luthier/Tooling/MachineFunctionEntryPoint.h"
+#include "luthier/Tooling/EntryPoint.h"
 #include "luthier/Tooling/MemoryAllocationAccessor.h"
 #include "luthier/Tooling/MetadataParserAnalysis.h"
 #include "luthier/Tooling/MockAMDGPULoader.h"
@@ -210,8 +210,6 @@ llvmGetPassPluginInfo() {
         [](llvm::MachineFunctionAnalysisManager &MFAM) {
           MFAM.registerPass(
               []() { return luthier::InstructionTracesAnalysis(); });
-          MFAM.registerPass(
-              []() { return luthier::MachineFunctionEntryPoint(); });
         });
 
     PB.registerPipelineParsingCallback(
